+++ scripts/install_core_dependencies.sh
@@ -19,8 +19,6 @@ sudo apt-get install -y libssl-dev libavahi-client-dev libavahi-common-dev libai
 sudo apt-get install -y libtool libudev1 libusb-1.0-0 libusb-1.0-0-dev libusb-dev
 sudo apt-get install -y libsamplerate-dev
 
-
-#
 # Ubuntu 20.04 LTS Desktop accomodations
 #
 # Following need to occur earlier in build
@@ -29,47 +27,90 @@ sudo apt-get install -y doxygen zlib1g zlib1g-dev libpng-dev libfaad-dev libfaac
 sudo apt-get install -y libwxgtk3.0-gtk3-dev libgps-dev
 
 # Qt5 Base Packages
-sudo apt-get install -y qtbase5-dev qtchooser qt5-qmake qtbase5-dev-tools
+sudo apt-get install -y qtbase5-dev 
+sudo apt-get install -y qtchooser
+sudo apt-get install -y qt5-qmake
+sudo apt-get install -y qtbase5-dev-tools
 
 # Python Baseline
-sudo apt-get install -y python3-pip python3-cairocffi python3-click python3-click-plugins python3-cycler python3-gdal python3-gi-cairo python3-kiwisolver python3-lxml python3-mako \
-python3-matplotlib python3-networkx python3-nose python3-numpy python3-opengl python3-pydot python3-pygraphviz python-qt4 qt5-default python3-pyqt5 python3-pyqt5.qtopengl python-pyside \
-python3-pyqtgraph python3-scapy python3-scipy python3-setuptools python3-sphinx python3-tk python3-tk-dbg python3-tornado python3-yaml python3-zmq python-cycler-doc \
-python-docutils python-gobject python-matplotlib-data python-matplotlib-doc python-networkx-doc 
-
+sudo apt-get install -y python3-pip 
+sudo apt-get install -y python3-cairocffi
+sudo apt-get install -y python3-click
+sudo apt-get install -y python3-click-plugins
+sudo apt-get install -y python3-cycler
+sudo apt-get install -y python3-gdal
+sudo apt-get install -y python3-gi-cairo
+sudo apt-get install -y python3-kiwisolver
+sudo apt-get install -y python3-lxml 
+sudo apt-get install -y python3-mako 
+sudo apt-get install -y python3-matplotlib
+sudo apt-get install -y python3-networkx
+sudo apt-get install -y python3-nose
+sudo apt-get install -y python3-numpy
+sudo apt-get install -y python3-opengl
+sudo apt-get install -y python3-pydot
+sudo apt-get install -y python3-pygraphviz
+sudo apt-get install -y python-qt4
+sudo apt-get install -y qt5-default
+sudo apt-get install -y python3-pyqt5
+sudo apt-get install -y python3-pyqt5.qtopengl
+sudo apt-get install -y python-pyside
+sudo apt-get install -y python3-pyqtgraph
+sudo apt-get install -y python3-scapy
+sudo apt-get install -y python3-scipy
+sudo apt-get install -y python3-setuptools
+sudo apt-get install -y python3-sphinx
+sudo apt-get install -y python3-tk
+sudo apt-get install -y python3-tk-dbg
+sudo apt-get install -y python3-tornado
+sudo apt-get install -y python3-yaml
+sudo apt-get install -y python3-zmq
+sudo apt-get install -y python-cycler-doc 
+sudo apt-get install -y python-docutils
+sudo apt-get install -y python-gobject
+sudo apt-get install -y python-matplotlib-data
+sudo apt-get install -y python-matplotlib-doc
+sudo apt-get install -y python-networkx-doc 
 sudo apt-get install -y python3-dbg
-
 sudo pip3 install pyinstaller
 sudo pip3 install pygccxml
 sudo pip3 install qtawesome
 
-# RTLSDR
-sudo apt-get install -y libudev1 libusb-1.0-0 libusb-1.0-0-dev libusb-dev
+
+# SDRdevice Baseline
+
+## RTLSDR
+sudo apt-get install -y libudev1 
+sudo apt-get install -y libusb-1.0-0 
+sudo apt-get install -y libusb-1.0-0-dev
+sudo apt-get install -y libusb-dev
 sudo pip3 install pyrtlsdr
-
-# HackRF
-sudo apt-get install -y libusb-1.0-0-dev libfftw3-dev
-
-# PlutoSDR
-sudo apt-get install -y libaio-dev libusb-1.0-0-dev 
-sudo apt-get install -y libserialport-dev libavahi-client-dev 
+## HackRF
+sudo apt-get install -y libusb-1.0-0-dev 
+sudo apt-get install -y libfftw3-dev
+## PlutoSDR
+sudo apt-get install -y libaio-dev 
+sudo apt-get install -y libusb-1.0-0-dev 
+sudo apt-get install -y libserialport-dev
+sudo apt-get install -y libavahi-client-dev 
 sudo apt-get install -y libxml2-dev bison flex libcdk5-dev 
-
-# LimeSDR
+## LimeSDR
 sudo apt-get install -y swig
 sudo apt-get install -y libsqlite3-dev
 sudo apt-get install -y libi2c-dev
 sudo apt-get install -y libusb-1.0-0-dev
 sudo apt-get install -y libwxgtk3.0-dev
 sudo apt-get install -y freeglut3-dev
+sudo apt-get install -y liboctave-dev
+sudo apt-get install -y libfltk1.3-dev
 
-# SoapySDR 
+## SoapySDR 
 sudo apt-get install -y swig
 sudo apt-get install -y avahi-daemon
 sudo apt-get install -y libavahi-client-dev
 sudo apt-get install -y libusb-1.0-0-dev
-sudo apt-get install -y python-dev python3-dev
-
+sudo apt-get install -y python-dev
+sudo apt-get install -y python3-dev
 
 echo -e "${SIGPI_BANNER_COLOR}"
 echo -e "${SIGPI_BANNER_COLOR} ##   Core Dependencies Installed"